feat(typescript): Generate typedefs, use TS
It turns out that sanity's code environment seems to support typescript
out of the box, so change all of the schema files to their TS
equivalents.

Also configure the sanity-codegen project to output typedefinitions that
we can use in other modules. Since the codegen is currently limited and
will not produce types for plugins we need to define some of the types
manually as overrides in the index.d.ts file.